+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+1.3.1
+-----
+
+Add ability to pass options to RTCPeerConnection and custom sdp transformer.
+
+1.3.0
+-----
+
+No longer forcibly close peers when network settings change. Clear packages
+upon connection instead.
+
 1.2.0
 -----